requirements/base.txt
Django==5.0
djangorestframework
psycopg2-binary
django-environ
redis
django-redis
celery
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
structlog
pytest
pytest-django
Pillow
